\name{avaliableIDs}
\alias{avaliableIDs}
\docType{data}
\title{Identifiers available in allSameID() function}
\description{
    avaliableIDs contains the different IDs that can be used in
    allSameID() function. It could be used before allSameID() function to know
    what ID can be used and how to write them.
}
\usage{data(avaliableIDs)}
\format{
    character vector
}
\keyword{Identifiers}
