\name{avaliableOrganism}
\alias{avaliableOrganism}
\docType{data}
\title{Organism available in allSameID() function}
\description{
    avaliableOrganism contains the differents organism that can be used in
    allSameID() function. It could be used before allSameID() function to know
    what organism can be used and how to write them.
}
\usage{data(avaliableOrganism)}
\format{
    character vector
}
\keyword{Organisms}
