\name{DExMAdata-package}
\alias{DExMAdata-package}
\alias{DExMAdata}
\docType{package}
\title{
    Data for DExMA package
}
\description{
Data objects needed to allSameID() function of DExMA package. There are also some objects that are necessary to be able to apply the examples of the DExMA package, which illustrate package functionality
}
\author{
Juan Antonio Villatoro-García and Pedro Carmona-Sáez

Maintainer: Juan Antonio Villatoro García<juan.villatoro@genyo.es>
}

\examples{
data(avaliableIDs)
data(SynonymsDExMA)
}

\keyword{datasets}
