\name{IDsDExMA}
\alias{IDsDExMA}
\docType{data}
\title{Gene/probe ID information}
\description{
    IDs contains the equivalent offical gene symbol to the different gene/id
    probes. It is used by allSameID() function to set all the genes/probes of
    the datasets in the same official gene symbol.
}
\usage{data(IDsDExMA)}
\format{
    list. Elements of the list are data.frames
}
\source{
    the information needed to make the object was downloaded from Gene
    Expression Omnibus (GEO) database and from
    ftp://ftp.ncbi.nih.gov/gene/DATA/gene_info.gz
}
\keyword{Identifiers}
