\name{SynonymsDExMA}
\alias{SynonymsDExMA}
\docType{data}
\title{Other names for identifiers information}
\description{
    Synonyms contains other possible gene names and the organisms they refer to.
    It is used by allSameID() function to set all the genes/probes of
    the datasets in the same official gene symbol.
}
\usage{data(SynonymsDExMA)}
\format{
    data.frame
}
\source{
    the information needed to make the object was downloaded from
    ftp://ftp.ncbi.nih.gov/gene/DATA/gene_info.gz
}
\keyword{Identifiers}
